lib/gg_components/automate-ventilation/requirements.txt
colorzero==2.0
gpiozero==2.0.1
pigpio==1.78
awsiotsdk